backend/requirements.txt
fastapi
uvicorn
python-multipart
fitz
pytesseract
Pillow